test(office): improve skill slug update coverage
